addressed cameron's visualization edits
