+++ src/articles/2022-11-07-reddit.mdx
@@ -1,6 +1,6 @@
 ---
-date: "2022-11-07"
-title: "Future Reddit article"
+date: "2022-11-09"
+title: "We scraped Berkeley’s Reddit page r/berkeley"
 bylineName:
   - "Isabella Borkovic"
   - "Erica Jean"
@@ -9,10 +9,10 @@ bylineUrl:
   - "https://www.dailycal.org/author/iborkovic"
   - "https://www.dailycal.org/author/ejean"
   - "https://www.dailycal.org/author/eagarwal"
-subhead: "Future Reddit subhead"
+subhead: "Here is what we found"
 featuredImage: "../images/HSI-Alison-Xiong-staff.png"
 imageAttribution: "Alison Xiong | Staff"
-aboutStory: "Data for this project come from information webscraped from the r/Berkeley subreddit."
+aboutStory: "Data for this project come from information webscraped from the r/Berkeley subreddit on Oct. 10, 2022."
 ---
 
 import { getImage, GatsbyImage } from "gatsby-plugin-image"
@@ -47,7 +47,7 @@ Parallel to the number of members, the total number of daily comments model a si
 
 <br />
 
-An analysis of posts reveals 9 p.m. to be the most popular time to post. This may suggest r/berkeley visitors are more likely to visit the page in the evenings –, when the workday is over and few classes are in session. However, 5 a.m. is the least popular time to post. Presumably, most users are asleep in the early morning.
+An analysis of posts reveals 9 p.m. to be the most popular time to post. This may suggest r/berkeley visitors are more likely to visit the page in the evenings, when the workday is over and few classes are in session. However, 5 a.m. is the least popular time to post. Presumably, most users are asleep in the early morning.
 
 <br />
 
@@ -72,4 +72,4 @@ _Isabella Borkovic is the deputy projects editor. Contact her at [iborkovic@dail
 
 _Erica Jean is a projects developer. Contact her at [ejean@dailycal.org](mailto:ejean@dailycal.org)_
 
-_Ekansh Agrawal is a projects developer. Contact him at [eagrawal@dailycal.orgeagrawal@dailycal.org)_
+_Ekansh Agrawal is a projects developer. Contact him at [eagrawal@dailycal.org](mailto:eagrawal@dailycal.org)_
