+++ src/visuals/reddit-visuals/postsPerFlairBarChart.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import {
   BarChart,
   ResponsiveContainer,
@@ -9,29 +9,6 @@ import {
   Tooltip,
 } from 'recharts';
 import data from './postsPerFlairData';
-
-class CustomizedAxisTick extends Component {
-  render() {
-    const {
-      x, y, payload,
-    } = this.props;
-
-    return (
-      <g transform={`translate(${x},${y})`}>
-        <text
-          x={0}
-          y={10}
-          dy={0}
-          textAnchor="end"
-          fill="#666"
-          transform="rotate(-35)"
-        >
-          {payload.value}
-        </text>
-      </g>
-    );
-  }
-}
 
 const PostsPerFlairBarChart = () => (
 
@@ -50,7 +27,7 @@ const PostsPerFlairBarChart = () => (
     <ResponsiveContainer height={680}>
       <BarChart width={400} height={680} data={data}>
         <CartesianGrid strokeDasharray="3 3" />
-        <XAxis dataKey="flair" tick={<CustomizedAxisTick />} height={125} />
+        <XAxis dataKey="flair" />
         <YAxis />
         <Tooltip />
         <Bar dataKey="Number of posts" fill="#fcba64" />
