+++ src/visuals/reddit-visuals/postsPerFlairData.js
@@ -3,11 +3,7 @@ const data = [
   { flair: 'CS/EECS', 'Number of posts': 5680 },
   { flair: 'Other', 'Number of posts': 5093 },
   { flair: 'Local', 'Number of posts': 1326 },
-  { flair: 'Events/Organizations', 'Number of posts': 1074 },
-  { flair: 'News', 'Number of posts': 422 },
-  { flair: 'Meta', 'Number of posts': 309 },
   { flair: 'Politics', 'Number of posts': 238 },
-  { flair: 'University faculty/staff', 'Number of posts': 45 },
 ];
 
 export default data;
