+++ src/visuals/reddit-visuals/postsPerHourBarChart.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { Component } from 'react';
 import {
   BarChart,
   ResponsiveContainer,
@@ -9,6 +9,29 @@ import {
   Bar,
 } from 'recharts';
 import data from './postsPerHourData';
+
+class CustomizedAxisTick extends Component {
+  render() {
+    const {
+      x, y, payload,
+    } = this.props;
+
+    return (
+      <g transform={`translate(${x},${y})`}>
+        <text
+          x={0}
+          y={10}
+          dy={0}
+          textAnchor="end"
+          fill="#666"
+          transform="rotate(-35)"
+        >
+          {payload.value}
+        </text>
+      </g>
+    );
+  }
+}
 
 const PostsPerHourBarChart = () => (
   <div>
@@ -26,7 +49,7 @@ const PostsPerHourBarChart = () => (
     <ResponsiveContainer height={600}>
       <BarChart width={400} height={600} data={data}>
         <CartesianGrid strokeDasharray="3 3" />
-        <XAxis dataKey="time" />
+        <XAxis dataKey="time" tick={<CustomizedAxisTick />} height={125} minTickGap={-10} />
         <YAxis />
         <Tooltip />
         <Bar dataKey="Number of posts" fill="#f38040" />
